cloud/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
supabase==2.11.0
python-dotenv==1.0.1
pydantic==2.10.0
python-jose[cryptography]==3.3.0
httpx==0.28.0
